pyproject.toml
[project]
name = "color-palette"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colour-science>=0.4.7",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "scipy>=1.17.1",
    "tqdm>=4.67.3",
]
